[TMail] refacor: Updating WireMockRagServerExtension to include scribe and refactoring it to a more generic name :WireMockAiServerExtension
