+++ tmail-backend/tmail-third-party/ai-bot/src/main/java/com/linagora/tmail/scribe/ChatCompletionResult.java
@@ -16,7 +16,7 @@
  *  more details.                                                   *
  ********************************************************************/
 
-package com.linagora.tmail.mailet.rag.httpclient;
+package com.linagora.tmail.scribe;
 
 import io.netty.handler.codec.http.HttpHeaders;
